requirements.txt
asgiref==3.5.2
certifi==2022.6.15
distlib==0.3.4
Django==4.0.6
django-ckeditor==6.4.2
django-js-asset==2.0.0
filelock==3.7.1
gunicorn==20.1.0
Pillow==9.2.0
pipenv==2022.7.4
platformdirs==2.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.15.1
virtualenv-clone==0.5.7
whitenoise==6.2.0
django-rosetta==0.9.8
django-parler==2.3
